General refinement
Added all art assets and changed the size of the view port from 1920 x 1080 to 800x600 and resized all the affected scenes which also involved reprogramming some parts of the name entry scene. Also re sized and changed some of the sizes of the cats. Furthermore I also ported the project from godot 3.5 to godot 4.2 so that the project can be exported to an ARM64 executable with relative ease.